Mengupdate views dan routes

+++ routes/web.php
@@ -21,19 +21,22 @@ Route::get('search', [SearchController::class, 'index'])->name('search');
 
 Route::view('login', 'auth.login')->name('login.form'); 
 Route::post('login', [AuthController::class, 'login'])->name('login');
+Route::get('categories', [CategoryController::class, 'index'])->name('categories.index');
 Route::get('categories/create', [CategoryController::class,'create'])->name('categories.create');
-Route::resource('categories', CategoryController::class);
+Route::get('categories/{category}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
+Route::get('categories/{category}', [CategoryController::class, 'show'])->name('categories.show');
+
 Route::prefix('api')->middleware('jwt.auth')->group(function () {
-
-    Route::resource('categories', CategoryApiController::class)->except('index','create','edit','show');
-
+    Route::post('categories', [CategoryApiController::class, 'store'])->name('categories.store');
+    Route::put('categories/{category}', [CategoryApiController::class, 'update'])->name('categories.update');
+    Route::delete('categories/{category}', [CategoryApiController::class, 'destroy'])->name('categories.destroy');
 }); 
 
-Route::prefix('api')->as('api.')->group(function () {
-    Route::post('/courses', [CourseApiController::class, 'store'])->name('courses.store');  
-    Route::put('/courses/{courseId}', [CourseApiController::class, 'update'])->name('courses.update');  
-    Route::delete('/courses/{courseId}', [CourseApiController::class, 'destroy'])->name('courses.destroy');
-    Route::get('/courses/{category_id?}', [CourseApiController::class, 'index'])->name('courses.index');
+Route::prefix('api')->middleware('jwt.auth')->group(function () {
+    Route::post('/courses', [CourseApiController::class, 'store'])->name('admin.courses.store');  
+    Route::put('/courses/{courseId}', [CourseApiController::class, 'update'])->name('admin.courses.update');  
+    Route::delete('/courses/{courseId}', [CourseApiController::class, 'destroy'])->name('admin.courses.destroy');
+    Route::get('/courses/{category_id?}', [CourseApiController::class, 'index'])->name('admin.courses.index');
 });
 
     Route::get('/courses/{category_id?}', [CourseController::class, 'index'])->name('admin.courses.index');
@@ -41,6 +44,7 @@ Route::prefix('api')->as('api.')->group(function () {
     Route::get('/courses/{id}/show', [CourseController::class, 'show'])->name('admin.courses.show');
     Route::get('/courses/{id}/edit', [CourseController::class, 'edit'])->name('admin.courses.edit');
 
+    Route::middleware('auth:api')->post('logout', [AuthController::class, 'logout'])->name('logout');
